regular commit

+++ src/get_map.c
@@ -82,7 +82,7 @@ int		get_map(char *s)
 	if (get_length(s) == -1)
 		return (-1);
 	read_map(s);
-	POSX = MW / 2.0 * (double)WALL;
-	POSY = MH / 2.0 * (double)WALL;
+	POSX = (double)((double)MW / 2.0 * (double)WALL);
+	POSY = (double)((double)MH / 2.0 * (double)WALL);
 	return (0);
 }
